pyproject.toml
[project]
name = "agent-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Gabriel Altay", email = "gabriel.altay@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "litellm>=1.67.0",
    "openai>=1.75.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "ruff>=0.11.6",
]

[project.scripts]
agent-examples = "agent_examples:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".venv",
    "__pypackages__",
    "build",
    "dist",
    "env_files",
]

# Same as Black.
line-length = 88

# Assume Python 3.12
target-version = "py312"

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "I"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
